# internal/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Message Example: invalid email or password
	Message string `json:"message"`
}

// MeResponse defines model for MeResponse.
type MeResponse struct {
	// UserId Example: 018f1234-5678-7abc-8def-0123456789ab
	UserId string `json:"user_id"`
}

// SigninRequest defines model for SigninRequest.
type SigninRequest struct {
	// Email Example: user@example.com
	Email openapi_types.Email `json:"email"`

	// Password Example: password123
	Password string `json:"password"`
}

// SigninResponse defines model for SigninResponse.
type SigninResponse struct {
	// UserId Example: 018f1234-5678-7abc-8def-0123456789ab
	UserId string `json:"user_id"`
}

// MeParams defines parameters for Me.
type MeParams struct {
	Session *string `form:"session,omitempty" json:"session,omitempty"`
}

// SignoutParams defines parameters for Signout.
type SignoutParams struct {
	Session *string `form:"session,omitempty" json:"session,omitempty"`
}

// SigninJSONRequestBody defines body for Signin for application/json ContentType.
type SigninJSONRequestBody = SigninRequest